use sharding table rule config to replace auto table rule config.
