+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-spring/shardingsphere-sharding-spring-boot-starter/src/main/java/org/apache/shardingsphere/sharding/spring/boot/ShardingRuleSpringBootConfiguration.java
@@ -20,6 +20,9 @@ package org.apache.shardingsphere.sharding.spring.boot;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.sharding.algorithm.config.AlgorithmProvidedShardingRuleConfiguration;
+import org.apache.shardingsphere.sharding.api.config.rule.ShardingTableRuleConfiguration;
+import org.apache.shardingsphere.sharding.api.config.strategy.sharding.ShardingStrategyConfiguration;
+import org.apache.shardingsphere.sharding.api.sharding.ShardingAutoTableAlgorithm;
 import org.apache.shardingsphere.sharding.spi.KeyGenerateAlgorithm;
 import org.apache.shardingsphere.sharding.spi.ShardingAlgorithm;
 import org.apache.shardingsphere.sharding.spring.boot.algorithm.KeyGenerateAlgorithmProvidedBeanRegistry;
@@ -27,7 +30,11 @@ import org.apache.shardingsphere.sharding.spring.boot.algorithm.ShardingAlgorith
 import org.apache.shardingsphere.sharding.spring.boot.condition.ShardingSpringBootCondition;
 import org.apache.shardingsphere.sharding.spring.boot.rule.YamlShardingRuleSpringBootConfiguration;
 import org.apache.shardingsphere.sharding.yaml.config.YamlShardingRuleConfiguration;
+import org.apache.shardingsphere.sharding.yaml.config.rule.YamlShardingAutoTableRuleConfiguration;
+import org.apache.shardingsphere.sharding.yaml.config.strategy.sharding.YamlShardingStrategyConfiguration;
 import org.apache.shardingsphere.sharding.yaml.swapper.ShardingRuleAlgorithmProviderConfigurationYamlSwapper;
+import org.apache.shardingsphere.sharding.yaml.swapper.rule.ShardingAutoTableRuleConfigurationYamlSwapper;
+import org.apache.shardingsphere.sharding.yaml.swapper.strategy.ShardingStrategyConfigurationYamlSwapper;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -36,7 +43,9 @@ import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 
+import java.util.Collection;
 import java.util.Collections;
+import java.util.LinkedList;
 import java.util.Map;
 import java.util.Optional;
 
@@ -52,6 +61,8 @@ public class ShardingRuleSpringBootConfiguration {
     
     private final ShardingRuleAlgorithmProviderConfigurationYamlSwapper swapper = new ShardingRuleAlgorithmProviderConfigurationYamlSwapper();
     
+    private final ShardingAutoTableRuleConfigurationYamlSwapper autoTableSwapper = new ShardingAutoTableRuleConfigurationYamlSwapper();
+    
     private final YamlShardingRuleSpringBootConfiguration yamlConfig;
     
     /**
@@ -66,7 +77,11 @@ public class ShardingRuleSpringBootConfiguration {
                                                        final ObjectProvider<Map<String, KeyGenerateAlgorithm>> keyGenerateAlgorithmProvider) {
         Map<String, ShardingAlgorithm> shardingAlgorithmMap = Optional.ofNullable(shardingAlgorithmProvider.getIfAvailable()).orElse(Collections.emptyMap());
         Map<String, KeyGenerateAlgorithm> keyGenerateAlgorithmMap = Optional.ofNullable(keyGenerateAlgorithmProvider.getIfAvailable()).orElse(Collections.emptyMap());
-        return swapper.swapToYamlConfigurationWithAlgorithms(yamlConfig.getSharding(), shardingAlgorithmMap, keyGenerateAlgorithmMap);
+        AlgorithmProvidedShardingRuleConfiguration result = swapper.swapToObject(yamlConfig.getSharding());
+        result.getTables().addAll(convertYamlAutoTablesToShardingTableRules(yamlConfig.getSharding(), shardingAlgorithmMap));
+        result.setShardingAlgorithms(shardingAlgorithmMap);
+        result.setKeyGenerators(keyGenerateAlgorithmMap);
+        return result;
     }
     
     /**
@@ -89,5 +104,29 @@ public class ShardingRuleSpringBootConfiguration {
     @Bean
     public static KeyGenerateAlgorithmProvidedBeanRegistry keyGenerateAlgorithmProvidedBeanRegistry(final Environment environment) {
         return new KeyGenerateAlgorithmProvidedBeanRegistry(environment);
+    }
+    
+    private Collection<ShardingTableRuleConfiguration> convertYamlAutoTablesToShardingTableRules(final YamlShardingRuleConfiguration yamlConfig, final Map<String, ShardingAlgorithm> algorithms) {
+        if (yamlConfig.getAutoTables().isEmpty()) {
+            return Collections.emptyList();
+        }
+        Collection<ShardingTableRuleConfiguration> result = new LinkedList<>();
+        for (Map.Entry<String, YamlShardingAutoTableRuleConfiguration> entry : yamlConfig.getAutoTables().entrySet()) {
+            YamlShardingAutoTableRuleConfiguration tableRuleConfig = entry.getValue();
+            tableRuleConfig.setLogicTable(entry.getKey());
+            result.add(autoTableSwapper.swapToObject(tableRuleConfig, getShardingCount(entry.getValue().getShardingStrategy(), algorithms)));
+        }
+        return result;
+    }
+    
+    private int getShardingCount(final YamlShardingStrategyConfiguration shardingStrategy, final Map<String, ShardingAlgorithm> shardingAlgorithmMap) {
+        ShardingStrategyConfiguration strategyConfiguration = new ShardingStrategyConfigurationYamlSwapper().swapToObject(shardingStrategy);
+        if (shardingAlgorithmMap.containsKey(strategyConfiguration.getShardingAlgorithmName())) {
+            ShardingAlgorithm shardingAlgorithm = shardingAlgorithmMap.get(strategyConfiguration.getShardingAlgorithmName());
+            if (shardingAlgorithm instanceof ShardingAutoTableAlgorithm) {
+                return ((ShardingAutoTableAlgorithm) shardingAlgorithm).getAutoTablesAmount();
+            }
+        }
+        return 1;
     }
 }
